Service Classes - ConnectionService
The Connection Service will be used to determine the connection string to our database at runtime. We use a local connection string during development. When we publish our app will need to grab the connection string from an Environment Variable. Environment Variables are stored with the hosting vendor. i.e. Heroku or Railway